pyproject.toml
[tool.poetry]
name = "machine-learning"
version = "0.1.0"
description = ""
authors = ["Alisson <alissonpintor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
streamlit = "^1.38.0"
requests = "^2.32.3"
scikit-learn = "^1.5.2"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scikit-fuzzy = "^0.5.0"
networkx = "^3.3"
cx-oracle = "^8.3.0"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
